+++ Video.js
@@ -229,6 +229,12 @@ export default class Video extends Component {
     }
   };
 
+  _onExternalPauseToggled = (event) => {
+    if (this.props.onExternalPauseToggled) {
+      this.props.onExternalPauseToggled(event.nativeEvent)
+    }
+  }
+
   getViewManagerConfig = viewManagerName => {
     if (!NativeModules.UIManager.getViewManagerConfig) {
       return NativeModules.UIManager[viewManagerName];
@@ -303,6 +309,7 @@ export default class Video extends Component {
       onAudioBecomingNoisy: this._onAudioBecomingNoisy,
       onPictureInPictureStatusChanged: this._onPictureInPictureStatusChanged,
       onRestoreUserInterfaceForPictureInPictureStop: this._onRestoreUserInterfaceForPictureInPictureStop,
+      onExternalPauseToggled: this._onExternalPauseToggled
     });
 
     const posterStyle = {
@@ -464,6 +471,7 @@ Video.propTypes = {
   onPictureInPictureStatusChanged: PropTypes.func,
   needsToRestoreUserInterfaceForPictureInPictureStop: PropTypes.func,
   onExternalPlaybackChange: PropTypes.func,
+  onExternalPauseToggled: PropTypes.func,
 
   /* Required by react-native */
   scaleX: PropTypes.number,